refactor webview component and handle client key and secret changed scenario
